+++ astropy/time/core.py
@@ -703,6 +703,50 @@ class TimeDelta(Time):
         if not isinstance(val, self.__class__):
             self._init_from_vals(val, val2, format, 'tai', copy)
 
+    def __neg__(self):
+        """Negation of a `TimeDelta` object."""
+        new = self.copy()
+        new._time.jd1 = -self._time.jd1
+        new._time.jd2 = -self._time.jd2
+        return new
+
+    def __abs__(self):
+        """Absolute value of a `TimeDelta` object."""
+        #TODO: for safety, ensure round-off errors are taken into account
+        jd1, jd2 = self._time.jd1, self._time.jd2
+        negative = jd1 + jd2 < 0
+        new = self.copy()
+        new._time.jd1 = np.where(negative, -jd1, jd1)
+        new._time.jd2 = np.where(negative, -jd2, jd2)
+        return new
+
+    def __mul__(self, other):
+        """Multiplication of `TimeDelta` objects by numbers/arrays."""
+        # check needed since otherwise the self.jd1 * other multiplication
+        # would enter here again (via __rmul__)
+        if isinstance(other, Time):
+            raise OperandTypeError(self, other)
+        #TODO: ensure round-off errors in jd1 are taken into account!!!
+        jd1 = self.jd1 * other
+        jd2 = self.jd2 * other
+        out = TimeDelta(jd1, jd2, format='jd')
+        if self.format != 'jd':
+            out = out.replicate(format=self.format)
+        return out
+
+    def __rmul__(self, other):
+        """Multiplication of numbers/arrays with `TimeDelta` objects."""
+        return self.__mul__(other)
+
+    def __div__(self, other):
+        """Division of `TimeDelta` objects by numbers/arrays."""
+        #TODO: ensure round-off errors in jd1 are taken into account!!!
+        return self.__mul__(1. / other)
+
+    def __truediv__(self, other):
+        """Division of `TimeDelta` objects by numbers/arrays."""
+        return self.__mul__(1. / other)
+
 
 class TimeFormat(object):
     """